+++ changelog.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Export wanted diagram as svg.
 - Add properties `categories` and `tags` in `ComponentDefinition`.
 - Add property `tags` in `ComponentAttributeDefinition`.
+- Add method `getTemplateData` in `ComponentRenderer` and `LinkRenderer`.
+- Add methods `initComponentRenderer` and `initLinkRenderer` in `DefaultDrawer`.
 
 ## [0.24.1] - 2024/06/18